LayoutTests/http/tests/security/contentSecurityPolicy/script-loads-with-img-src-expected.txt

CONSOLE MESSAGE: line 1: Unrecognized Content-Security-Policy directive 'script-img'.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
